cheesy data
OK, I looked at doing this with ActiveRecord, but I wanted it done as
quickly as possible. moving this to ARec is probably the way to go - in
fact part of me just wants to replace it with a Rails app, so I get the
console, but you can get the console so easy that it doesnt' matter.

anyway, for now, it's a flat file. easy enough to change it later, but
YAGNI.